// resources/js/components/types/types.ts
export interface SearchParams {
    search?: string;
    filters?: string[];
    sort?: string;
    direction?: 'asc' | 'desc';
    page?: number;
    per_page?: number;
    [key: string]: any;
}
// ...
export interface DebouncedSearchReturn {
    params: SearchParams;
    setParams: (params: SearchParams) => void;
    setTimeDebounce: (time: number) => void;
    isLoading: boolean;
}